+++ packages/cysync-core-constants/src/i18n/types.ts
@@ -361,6 +361,7 @@ interface LangSend {
       error: string;
     };
     rentExemptFeeWarning: string;
+    parentAssetInsufficient: string;
   };
   summary: {
     title: string;