+++ packages/cysync-core/src/dialogs/Send/Dialogs/Recipient.tsx
@@ -14,6 +14,7 @@ import {
   InformationIcon,
   LangDisplay,
   LeanBox,
+  MessageBox,
   ScrollableContainer,
   Typography,
   useTheme,
@@ -28,6 +29,7 @@ import logger from '~/utils/logger';
 import { AddressAndAmountSection, FeeSection } from './Components';
 
 import { useSendDialog } from '../context';
+import { CoinFamily, coinList } from '@cypherock/coins';
 
 export const Recipient: React.FC = () => {
   const {
@@ -39,11 +41,16 @@ export const Recipient: React.FC = () => {
     isAccountSelectionDisabled,
     isPreparingTxn,
     providerName,
+    selectedAccountParent,
+    getComputedFee,
   } = useSendDialog();
   const lang = useAppSelector(selectLanguage);
   const button = lang.strings.buttons;
   const theme = useTheme();
   const displayText = lang.strings.send.recipient;
+  const [showParentBalanceWarning, setShowParentBalanceWarning] =
+    useState(false);
+  const [minParentBalance, setMinParentBalance] = useState<string>('0');
 
   const getBalanceToDisplay = () => {
     const account = selectedAccount;
@@ -129,6 +136,31 @@ export const Recipient: React.FC = () => {
     initialize();
   }, []);
 
+  useEffect(() => {
+    if (selectedAccount && selectedAccountParent && transaction) {
+      const parentBalance = selectedAccountParent?.balance ?? '0';
+      const computedFee = getComputedFee(
+        selectedAccount.familyId as CoinFamily,
+        transaction,
+      );
+      setShowParentBalanceWarning(
+        new BigNumber(parentBalance).isLessThanOrEqualTo(
+          new BigNumber(computedFee),
+        ),
+      );
+      const { amount: minParsedAmount, unit } = getParsedAmount({
+        coinId: selectedAccountParent.parentAssetId,
+        assetId: selectedAccountParent.assetId,
+        unitAbbr: getDefaultUnit(
+          selectedAccountParent.parentAssetId,
+          selectedAccountParent.assetId,
+        ).abbr,
+        amount: computedFee,
+      });
+      setMinParentBalance(`${minParsedAmount} ${unit.abbr}`);
+    }
+  }, [selectedAccount, selectedAccountParent, transaction]);
+
   const handleSubmit = useCallback(() => {
     logger.info('Form Submit: Recipient', {
       source: `Send/${Recipient.name}`,
@@ -170,6 +202,19 @@ export const Recipient: React.FC = () => {
         )}
       </DialogBoxBody>
       <ScrollableContainer>
+        {showParentBalanceWarning && selectedAccountParent && (
+          <Container px={5}>
+            <MessageBox
+              type="warning"
+              text={lang.strings.send.recipient.parentAssetInsufficient}
+              variables={{
+                minParentBalance,
+                network: coinList[selectedAccountParent.parentAssetId].name,
+                unit: coinList[selectedAccountParent.parentAssetId].abbr,
+              }}
+            />
+          </Container>
+        )}
         <AddressAndAmountSection
           disableInputs={isAccountSelectionDisabled}
           providerName={providerName}