+++ packages/cysync-core/src/services/index.ts
@@ -1,2 +1,3 @@
 export * from './feedbackService';
 export * from './inheritance';
+export * from './analytics';